requirements.txt
beautifulsoup4==4.14.3
pandas==2.3.3
python-telegram-bot==22.7
requests==2.33.1
google-genai
